requirements.txt
requests
beautifulsoup4
html5lib
python-decouple